BE/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Balo_BkavZalo\BE\app\auth\auth.py: 2
PyJWT == 2.2.0

# D:\Balo_BkavZalo\BE\app\api.py: 4
# D:\Balo_BkavZalo\BE\app\controller\UserController.py: 5,6,7,8
# D:\Balo_BkavZalo\BE\app\db_connect.py: 4
# D:\Balo_BkavZalo\BE\main.py: 2
app == 0.0.1

# D:\Balo_BkavZalo\BE\app\api.py: 2
# D:\Balo_BkavZalo\BE\app\auth\auth.py: 4
# D:\Balo_BkavZalo\BE\app\controller\UserController.py: 1,2,3,4
fastapi == 0.70.0

# D:\Balo_BkavZalo\BE\app\db_connect.py: 1,2,3
firebase_admin == 5.0.3

# D:\Balo_BkavZalo\BE\app\auth\auth.py: 5
passlib == 1.7.4

# D:\Balo_BkavZalo\BE\app\auth\RegisterRequest.py: 1
# D:\Balo_BkavZalo\BE\app\model\Account.py: 1
pydantic == 1.8.2

# D:\Balo_BkavZalo\BE\main.py: 1
uvicorn == 0.15.0
